abstract React-Query into Custom Hooks

+++ src/services/apiCabins.js
@@ -11,21 +11,28 @@ export async function getCabins() {
 }
 
 export async function createEditCabin(newCabin, id) {
-  const hasImagePath = newCabin.image?.statsWith?.(supabaseUrl);
+  console.log("received from app: ", newCabin.image);
+  console.log(id);
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-
   const imagePath = hasImagePath
     ? newCabin.image
     : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
+  console.log(hasImagePath);
+  console.log("generate new path: ", imagePath);
+
+  // 1. Create/edit cabin
   let query = supabase.from("cabins");
 
+  // A) CREATE
   if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
+  // B) EDIT
   if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
 
   const { data, error } = await query.select().single();
@@ -35,14 +42,20 @@ export async function createEditCabin(newCabin, id) {
     throw new Error("Cabin could not be created");
   }
 
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
-    throw new Error("Cabin could not be created");
+    throw new Error(
+      "Cabin image could not be uploaded and the cabin was not created"
+    );
   }
 
   return data;
